refactor: update state taxes to latest code patterns
