--- src/components/Company/StateTaxes/StateTaxesForm/Actions.tsx
@@ -1,21 +1,0 @@
-import { useTranslation } from 'react-i18next'
-import { useStateTaxesForm } from './context'
-import { ActionsLayout } from '@/components/Common'
-import { useComponentContext } from '@/contexts/ComponentAdapter/useComponentContext'
-
-export function Actions() {
-  const { t } = useTranslation('Company.StateTaxes', { keyPrefix: 'form' })
-  const { handleCancel, isPending } = useStateTaxesForm()
-  const Components = useComponentContext()
-
-  return (
-    <ActionsLayout>
-      <Components.Button variant="secondary" onClick={handleCancel}>
-        {t('cancelCta')}
-      </Components.Button>
-      <Components.Button variant="primary" type="submit" isDisabled={isPending}>
-        {t('saveCta')}
-      </Components.Button>
-    </ActionsLayout>
-  )
-}
