--- src/components/Company/StateTaxes/StateTaxesList/List.tsx
@@ -1,56 +1,0 @@
-import { useTranslation } from 'react-i18next'
-import { useStateTaxesList } from './context'
-import { DataView, EmptyData, useDataView } from '@/components/Common'
-import type { STATES_ABBR } from '@/shared/constants'
-import { useComponentContext } from '@/contexts/ComponentAdapter/useComponentContext'
-
-export const List = () => {
-  const { stateTaxRequirements, handleChange } = useStateTaxesList()
-  const Components = useComponentContext()
-
-  const { t } = useTranslation('Company.StateTaxes', { keyPrefix: 'list' })
-  const { t: statesHash } = useTranslation('common', { keyPrefix: 'statesHash' })
-
-  const { ...dataViewProps } = useDataView({
-    data: stateTaxRequirements,
-    columns: [
-      {
-        key: 'state',
-        title: t('requirementsListCol1'),
-        render: requirement => {
-          return <span>{statesHash(requirement.state as (typeof STATES_ABBR)[number])}</span>
-        },
-      },
-      {
-        key: 'status',
-        title: t('requirementsListCol2'),
-        render: requirement => {
-          return (
-            <>
-              <Components.Badge status={requirement.setupComplete ? 'success' : 'warning'}>
-                {requirement.setupComplete ? t('completeBadge') : t('incompleteBadge')}
-              </Components.Badge>
-            </>
-          )
-        },
-      },
-    ],
-    itemMenu: requirement => {
-      return (
-        <Components.Button
-          variant="secondary"
-          onClick={() => {
-            if (requirement.state) handleChange(requirement.state)
-          }}
-        >
-          {requirement.setupComplete ? t('editStateTaxCta') : t('continueStateTaxSetupCta')}
-        </Components.Button>
-      )
-    },
-
-    emptyState: () => (
-      <EmptyData title={t('emptyTableTitle')} description={t('emptyTableDescription')} />
-    ),
-  })
-  return <DataView label={t('requirementsListLabel')} {...dataViewProps} />
-}
